Q&A Chatbot/requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
langchain-community
langchain-google-genai
google-generativeai